+++ client/src/shared/hooks/api/useQuizQueries.hook.tsx
@@ -1,6 +1,6 @@
 import { URLS } from "../../../configs/Links";
-import { useMutationFactory } from "./useMutationFactory";
-import { useQueryFactory } from "./useQueryFactory";
+import { useMutationFactory } from "./_useMutationFactory";
+import { useQueryFactory } from "./_useQueryFactory";
 
 /**
  * @description This hook is used to create hooks for quiz queries.
